test(Overlay): 补充Overlay组件的单元测试
